refactor(web): AddIntegrationModal becomes the host chassis; the four wizard bodies become platform fragments (S3 §10)
The install wizard was a 3794-line monolith in which every platform's pane,
probe, poll and validator lived in one component's state. Adding a platform
meant editing that file in a dozen places, and the platform name was core
knowledge: `platform === 'slack'` decided the footer, the identity chrome, the
transport default, the reuse input and the free-bot predicate.

This lands the wizard half of `WebPlatformModule` (the seam published in #573).
The chassis keeps exactly §10's list — picker tiles and their
`daemon.caps.platforms` gate, the existing/create mode cards, the generic
free-bot reuse list, the share toggle, the error banner, the footer, and the
webhook/github sections, which mint an inbound trigger rather than a bot
identity and are therefore CORE fragments, not modules. Everything per-platform
moved to `platforms/<id>/`, reachable only through `WizardHost`.

WHY THIS SHAPE

- Two publication channels, not two-way access. A fragment drives host-rendered
  chrome via `setFooter` / `setIdentityChrome` instead of reaching into modal
  internals. Both keep the CALLBACKS in a ref and only the RENDERED fields in
  state, so a Body may republish on every commit without looping the host — an
  unchanged publication returns the previous object and React bails out.
- The platform-switch reset seam disappears. The chassis mounts
  `<wizard.Body key={platform}>`, so a switch unmounts the old fragment and its
  state goes with it. The monolith's two hand-maintained 20-line `set*('')`
  reset lists are gone; they can no longer drift from the fields they reset.
- One deployment probe. `GET /slack/config` is Slack-NAMED but
  deployment-SCOPED, and the monolith called it from two effects. It is now one
  shared SWR key: the chassis reads `relayAvailable`/`relayPublicUrl` for
  `WizardHost.relayCapability`; the Slack-funnel flags on the same DTO stay
  Slack module state.

BEHAVIOR PRESERVED, DELIBERATELY

Every derived flag was moved arm-for-arm, including the awkward ones:

- `(!b.teamId || b.shareable)` applied to all platforms in the monolith but is
  Slack-only in truth — `Bot.teamId` is the Slack workspace id and no other
  platform ever writes it — so it became Slack's `freeBotFilter` and the other
  three add nothing.
- Telegram/Discord reuse spread `shareable` under `wantShared`, which was dead:
  `wantShared` required `platform === 'slack'`. Their `buildReuseInput` drops
  the spread and commits the identical payload.
- The busy-label swap (`'Creating…'` during auto-config, else `'Connecting…'`)
  is reproduced inside the Slack fragment's own publication.
- `nextTransport = transport ?? (relayAvailable ? 'http' : 'socket')` relied on
  the pick being NULLABLE. The host hands fragments a resolved value, so the
  Slack Body tracks `transportPicked` to keep the same three-way outcome.
- The one pre-existing `isMobile` fork (the walkthrough disclosure) moved
  verbatim into `wizard-chrome.tsx`. No new fork was introduced; the wizard
  stays one responsive tree of base + `desktop:` classes.

Mock mode rides on `WizardHost.mockMode`, and the monolith's Slack-only mock
branch became `MOCK_DEPLOYMENT_CONFIG` with the same four flags, so design/dev
with no CP still reaches every pane.

WHAT THE TESTS PIN

`platforms/reuse.test.ts` covers the pure half of the facet — `freeBotFilter`
and `buildReuseInput` per platform, plus `affordances` and the
region-parameterised copy. Nothing renders these, so an arm could drift
silently; the assertions state each platform's rule explicitly (Slack's
not-yet-shared workspace install, Feishu's region match and its legacy
region-less bot, the absent `shareable` key vs `shareable: false`) and pin the
registry's id set and picker order.

Verified additionally by extracting every multi-word string from the monolith
and from the refactored tree: no user-visible copy was dropped and none was
invented.
