+++ packages/web/src/components/console/platforms/telegram/privacy-auto-refresh.ts
